function that concatenates two string

+++ 0x0B-malloc_free/2-str_concat.c
@@ -10,7 +10,7 @@
 
 char *str_concat(char *s1, char *s2)
 {
-	unsigned short int a, b, size, i, j;
+	unsigned int a, b, size, i, j;
 	char *new_str;
 
 	if (s1 == NULL)
